requirements.txt
arrow==0.14.7
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.8
ics==0.7
idna==3.3
importlib-metadata==4.8.2
keyring==23.4.0
python-dateutil==2.8.2
pywin32-ctypes==0.2.0
PyYAML==6.0
requests==2.26.0
six==1.16.0
TatSu==5.6.1
todoist-api-python==1.1.0
urllib3==1.26.7
zipp==3.6.0
